python/blacklist/requirements.txt
coloredlogs
stormshield-sns-api
